Try fix proxy client get_or_build
