+++ src/shared_client.rs
@@ -9,13 +9,22 @@ use std::{fs::File, io::BufReader};
 use hyper_proxy::{Proxy, ProxyConnector, Intercept};
 use crate::config::AppConfig;
 use std::time::Duration;
-use crate::proxy::ClientOptions;
 use std::str::FromStr;
 use once_cell::sync::Lazy;
 use dashmap::DashMap;
 
 #[allow(dead_code)]
 static CLIENT_CACHE: Lazy<DashMap<ClientKey, Client<hyper_rustls::HttpsConnector<hyper::client::HttpConnector>>>> = Lazy::new(DashMap::new);
+static CLIENT_CACHE_PROXY: Lazy<DashMap<ClientKey, Client<ProxyConnector<HttpsConnector<HttpConnector>>>>> = Lazy::new(|| DashMap::new());
+
+#[derive(Clone, Debug, Eq, PartialEq, Hash)]
+pub struct ClientOptions {
+    pub use_proxy: bool,
+    pub proxy_addr: Option<String>,
+    pub use_cert: bool,
+    pub cert_path: Option<String>,
+    pub key_path: Option<String>,
+}
 
 #[derive(Clone, Hash, Eq, PartialEq, Debug)]
 pub struct ClientKey {
@@ -39,16 +48,41 @@ impl ClientKey {
     }
 }
 
-#[allow(dead_code)]
-pub fn get_or_build_client(opts: ClientOptions, state: Arc<AppConfig>) -> Client<hyper_rustls::HttpsConnector<hyper::client::HttpConnector>> {
+pub fn get_or_build_client(
+    opts: ClientOptions,
+    state: Arc<AppConfig>,
+) -> Client<HttpsConnector<HttpConnector>> {
     let key = ClientKey::from_options(&opts);
 
     if let Some(client) = CLIENT_CACHE.get(&key) {
         return client.clone();
     }
 
-    let client = build_hyper_client_cert(opts.clone(), &state);
+    let client = {
+        if opts.use_cert {
+            build_hyper_client_cert(opts.clone(), &state)
+        } else {
+            build_hyper_client_normal(&state)
+        }
+    };
+
     CLIENT_CACHE.insert(key, client.clone());
+    client
+}
+
+pub fn get_or_build_client_proxy(
+    opts: ClientOptions,
+    state: Arc<AppConfig>,
+) -> Client<ProxyConnector<HttpsConnector<HttpConnector>>> {
+    let key = ClientKey::from_options(&opts);
+
+    if let Some(client) = CLIENT_CACHE_PROXY.get(&key) {
+        return client.clone();
+    }
+
+    let client = build_hyper_client_proxy(opts.clone(), &state);
+
+    CLIENT_CACHE_PROXY.insert(key, client.clone());
     client
 }
